requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.41.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1
idna==2.10
oauthlib==3.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
psycopg2==2.8.6
django_compressor==2.4
django-libsass==0.8
Pillow